Broke location error reducer out into separate file
