+++ src/reducers/locations.js
@@ -1,7 +1,5 @@
 import {
-  GET_LOCATION_REQUEST,
   GET_LOCATION_RESPONSE,
-  GET_LOCATION_ERROR,
   GET_TAXONOMY_RESPONSE,
   OPTIMISTIC_UPDATE_LOCATION,
   ROLLBACK_UPDATE_LOCATION,
@@ -41,18 +39,6 @@ function constructNewStateWithUpdatedPhones(state, action, newPhones, location, 
     },
   };
 }
-
-export const locationErrorsReducer = (state = {}, action) => {
-  switch (action.type) {
-    case GET_LOCATION_REQUEST:
-      return { ...state, [action.locationId]: null };
-    case GET_LOCATION_ERROR:
-      return { ...state, [action.locationId]: action.errorMessage };
-    default:
-      break;
-  }
-  return state;
-};
 
 export const locationsReducer = (state = {}, action) => {
   const dateString = new Date().toISOString();
